requirements.txt
numpy
matplotlib
pandas
python-dotenv
colorama
